model: generator: Add dependency relationship between containers.
Some containers might be dependant on another. Container A depends on
container B if some actions that the player is able to perform on A
become available only after opening B. This relationship is transitive,
asymmetric and nonreflexive.

This information is a hint for Generator necessary for avoiding unsolvable
cycles. For example, in order to be able to do something with the false
bottom of the box, you have to open the box. So that tools for opening
the box mustn't be placed in the false bottom.